delete.phpに論理削除実装、loginしているユーザー情報の表示実装

+++ index.php
@@ -3,10 +3,13 @@
   require('db_connect.php');
 
   // ログインチェック
-  if (!isset($_SESSION['login_id'])) {
-    // ログインしていない時
-    header('Location: login.php');
-  } else {
+  // 前回ログインした時間より1時間以上経ってる時にアクセスしたらlogin.phpに遷移する
+  // 1時間 == 3600秒
+  if (isset($_SESSION['login_id']) && $_SESSION['time'] + 3600 > time()) {
+
+    // $_SESSION['time']の時間を更新
+    $_SESSION['time'] = time();
+
     // ログインしている時
     $sql = 'SELECT * FROM `members` WHERE `member_id`=?';
     $data = array($_SESSION['login_id']);
@@ -14,6 +17,9 @@
     $stmt->execute($data);
     // ログインしているユーザーのデータが入っている
     $login_member = $stmt->fetch(PDO::FETCH_ASSOC);
+  } else {
+    // ログインしていない時
+    header('Location: login.php');
   }
 
   // つぶやくボタンが押された時
@@ -78,7 +84,7 @@
   $start_page = ($page - 1) * $max_page_tweet;
 
   // 一覧用のつぶやき全件を最新順に取得
-  $tweet_sql = "SELECT `tweets`.*, `members`.`nickname`, `members`.`picture_path`, `members`.`created` AS `member_created` FROM `tweets` LEFT JOIN `members` ON `tweets`.`member_id`=`members`.`member_id` ORDER BY `tweets`.`created` DESC LIMIT ".$start_page.",".$max_page_tweet;
+  $tweet_sql = "SELECT `tweets`.*, `members`.`nickname`, `members`.`picture_path`, `members`.`created` AS `member_created` FROM `tweets` LEFT JOIN `members` ON `tweets`.`member_id`=`members`.`member_id` WHERE `delete_flag`=0 ORDER BY `tweets`.`created` DESC LIMIT ".$start_page.",".$max_page_tweet;
   // SELECT 取得したいカラム FROM 取得したいテーブル LEFT JOIN 繋げたいテーブル ON 取得したいテーブル.繋げるキー=繋げたいテーブル.繋げるキー ORDER BY 順番 昇順か降順か LIMIT 開始位置,取得する個数
   // LIMIT == 取得するデータの制限をする
   // LIMIT 開始位置, 取得する個数
@@ -97,10 +103,10 @@
     $tweets[] = $tweet;
   }
 
-  echo '<br>';
-  echo '<pre>';
-  var_dump($tweets);
-  echo '</pre>';
+  // echo '<br>';
+  // echo '<pre>';
+  // var_dump($tweets);
+  // echo '</pre>';
 ?>
 
 <!DOCTYPE html>
@@ -131,7 +137,7 @@
                   <span class="icon-bar"></span>
                   <span class="icon-bar"></span>
               </button>
-              <a class="navbar-brand" href="index.html"><span class="strong-title"><i class="fa fa-twitter-square"></i> Seed SNS</span></a>
+              <a class="navbar-brand" href="index.php"><span class="strong-title"><i class="fa fa-twitter-square"></i> Seed SNS</span></a>
           </div>
           <!-- Collect the nav links, forms, and other content for toggling -->
           <div class="collapse navbar-collapse" id="bs-example-navbar-collapse-1">
@@ -147,7 +153,7 @@
   <div class="container">
     <div class="row">
       <div class="col-md-4 content-margin-top">
-        <legend>ようこそ<?php echo $login_member['nickname']; ?>さん！</legend>
+        <legend>ようこそ<a href="login_user.php?member_id=<?php echo $_SESSION['login_id']; ?>"><?php echo $login_member['nickname']; ?></a>さん！</legend>
         <form method="post" action="" class="form-horizontal" role="form">
             <!-- つぶやき -->
             <div class="form-group">
@@ -159,27 +165,38 @@
           <ul class="paging">
             <input type="submit" class="btn btn-info" value="つぶやく">
                 &nbsp;&nbsp;&nbsp;&nbsp;
-                <li><a href="index.php?page=<?php echo $page - 1 ?>" class="btn btn-default">前</a></li>
+                <?php if ($page > 1): ?>
+                  <li><a href="index.php?page=<?php echo $page - 1 ?>" class="btn btn-default">前</a></li>
+                <?php else: ?>
+                  <li>前</li>
+                <?php endif ?>
                 &nbsp;&nbsp;|&nbsp;&nbsp;
-                <li><a href="index.php?page=<?php echo $page + 1 ?>" class="btn btn-default">次</a></li>
+                <?php if ($page < $all_pages_number): ?>
+                  <li><a href="index.php?page=<?php echo $page + 1 ?>" class="btn btn-default">次</a></li>
+                <?php else: ?>
+                 <li>次</li>
+                <?php endif ?>
+                 <li><?php echo $page; ?> / <?php echo $all_pages_number; ?> Page</li>
           </ul>
         </form>
       </div>
-      
+
       <div class="col-md-8 content-margin-top">
         <?php foreach ($tweets as $tweet) { ?>
         <div class="msg">
-          <img src="http://c85c7a.medialib.glogster.com/taniaarca/media/71/71c8671f98761a43f6f50a282e20f0b82bdb1f8c/blog-images-1349202732-fondo-steve-jobs-ipad.jpg" width="48" height="48">
+          <img src="picture_path/<?php echo $tweet['picture_path']; ?>" width="48" height="48">
           <p>
-            <?php echo $tweet['tweet']; ?><span class="name"><a href="profile.php?tweet_id=<?php echo $tweet['tweet_id']; ?>">(Seed Kun) </a></span>
+            <?php echo $tweet['tweet']; ?><span class="name"><a href="profile.php?tweet_id=<?php echo $tweet['tweet_id']; ?>">(<?php echo $tweet['nickname']; ?>) </a></span>
             [<a href="#">Re</a>]
           </p>
           <p class="day">
             <a href="view.php?tweet_id=<?php echo $tweet['tweet_id']; ?>">
               <?php echo $tweet['created']; ?>
             </a>
-            [<a href="edit.php?tweet_id=<?php echo $tweet['tweet_id']; ?>" style="color: #00994C;">編集</a>]
-            [<a href="delete.php?tweet_id=<?php echo $tweet['tweet_id']; ?>" style="color: #F33;">削除</a>]
+            <?php if ($tweet['member_id'] == $login_member['member_id']): ?>
+              [<a href="edit.php?tweet_id=<?php echo $tweet['tweet_id']; ?>" style="color: #00994C;">編集</a>]
+              [<a href="delete.php?tweet_id=<?php echo $tweet['tweet_id']; ?>" style="color: #F33;">削除</a>]
+            <?php endif ?>
           </p>
         </div>
         <?php } ?>
